requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
choicelib==0.1.5
colorama==0.4.6
cryptography==46.0.6
et_xmlfile==2.0.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
msgspec==0.19.0
multidict==6.7.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.33.0
selenium==4.41.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vbml==1.1.post1
vkbottle==4.7.0
vkbottle-types==5.199.99.12
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.23.0
